requirements.txt
boto3>=1.26.0
PyYAML>=6.0
colorama>=0.4.6
typing-extensions>=4.5.0